resolved github PR comments
