+++ apps/web/src/utils/article.ts
@@ -1,4 +1,4 @@
-import { Article } from "@/types/news.types";
+import { Article } from "@/types/articles.types";
 import { endpoints } from "@/data/endpoints";
 import { getDaysAgo } from "@/utils/date";
 
@@ -54,25 +54,4 @@ export const getCategoryLabel = (
     category?.slug,
   ].find((v) => typeof v === "string" && v.trim());
   return label ? label.trim() : "";
-};
-
-export const fetchCategories = async (
-  browseableArticles: { field: string }[],
-) => {
-  const endpoint = endpoints.categories.get;
-  try {
-    const res = await fetch(endpoint, {
-      headers: { Accept: "application/json" },
-    });
-    const result = await res.json().catch(() => ({}));
-    if (!res.ok || !result.success) throw new Error();
-    const categories = Array.isArray(result.data) ? result.data : [];
-    return categories;
-  } catch {
-    // Fallback: populate from article data
-    const fields = [
-      ...new Set(browseableArticles.map((a) => a.field).filter(Boolean)),
-    ].sort();
-    return fields;
-  }
 };
