+++ apps/web/src/utils/sanity.ts
@@ -1,5 +1,3 @@
-import { sanityQuery } from "@/actions/sanity";
-import type { Article } from "@/types/news.types";
 import { createClient } from "@sanity/client";
 import imageUrlBuilder from "@sanity/image-url";
 import { defineQuery } from "next-sanity";
@@ -34,37 +32,6 @@ export function estimateReadTime(text: string) {
   return readTime;
 }
 
-export async function fetchArticlesFromSantry(): Promise<Article[]> {
-  let articles: Article[] = [];
-
-  try {
-    const query = `
-      *[_type == "article" && status == "published"]
-      | order(publicationDate desc) {
-        "id": _id,
-        title,
-        "slug": slug.current,
-        author,
-        source,
-        "summary": excerpt,
-        field,
-        region,
-        readTime,
-        "imageUrl": coverImage.asset->url,
-        featured,
-        "date": publicationDate,
-        externalUrl,
-        sourceType
-      }
-    `;
-    articles = await sanityQuery(query);
-  } catch (err) {
-    console.error("Failed to load articles:", err);
-    throw new Error("Failed to load articles");
-  }
-
-  return articles;
-}
 export const POST_QUERY = defineQuery(
   `*[_type == "article" && slug.current == $slug][0]`,
 );
